test: move envMainRuntimeFile to fixtures helper
