+++ src/e2e-helper/e2e-fixtures-helper.ts
@@ -381,4 +381,40 @@ export default () => 'comp${index} and ' + ${nextComp}();`;
       cwd: destDir,
     });
   }
+
+  populateEnvMainRuntime(
+    filePathRelativeToLocalScope: string,
+    { envName, dependencies }: { envName: string; dependencies: any }
+  ): void {
+    const capitalizedEnvName = envName[0].toUpperCase() + envName.substring(1);
+    return this.fs.outputFile(
+      filePathRelativeToLocalScope,
+      `
+import { MainRuntime } from '@teambit/cli';
+import { ReactAspect, ReactMain } from '@teambit/react';
+import { EnvsAspect, EnvsMain } from '@teambit/envs';
+import { ${capitalizedEnvName}Aspect } from './${envName}.aspect';
+
+export class EnvMain {
+  static slots = [];
+
+  static dependencies = [ReactAspect, EnvsAspect];
+
+  static runtime = MainRuntime;
+
+  static async provider([react, envs]: [ReactMain, EnvsMain]) {
+    const templatesReactEnv = envs.compose(react.reactEnv, [
+      envs.override({
+      getDependencies: () => (${JSON.stringify(dependencies)}),
+      })
+    ]);
+    envs.registerEnv(templatesReactEnv);
+    return new EnvMain();
+  }
+}
+
+${capitalizedEnvName}Aspect.addRuntime(EnvMain);
+`
+    );
+  }
 }
